# RavenDB-25098 Optimize webpack build and update packages