--- a/src/Raven.Studio/tsconfig.json
+++ b/src/Raven.Studio/tsconfig.json
@@ -32,7 +32,6 @@
             "d3",
             "rbush",
             "requirejs",
-            "jquery.blockui",
             "prismjs",
             "toastr",
             "diff",