--- a/src/Raven.Studio/typescript/components/common/PopoverWithHover.tsx
+++ b/src/Raven.Studio/typescript/components/common/PopoverWithHover.tsx
@@ -44,7 +44,6 @@ export function PopoverWithHover(props: PopoverWithHoverProps) {
             if (!overElement.current) {
                 setOpen(false);
                 if (tooltipContext.currentTarget === div) {
-                    // eslint-disable-next-line react-compiler/react-compiler
                     tooltipContext.currentTarget = null;
                     tooltipContext.closeAction = null;
                 }