--- a/src/Raven.Studio/typescript/components/common/virtualTable/columnProviders/useDocumentColumnsProvider.tsx
+++ b/src/Raven.Studio/typescript/components/common/virtualTable/columnProviders/useDocumentColumnsProvider.tsx
@@ -37,7 +37,6 @@ export function useDocumentColumnsProvider(props: UseDocumentColumnsProviderProp
         if (hasCheckbox) {
             columnsDefs.push(columnCheckbox as ColumnDef<document>);
             initialColumnVisibility[columnCheckbox.header.toString()] = true;
-            // eslint-disable-next-line react-compiler/react-compiler
             availableWidth -= columnCheckbox.size;
         }