--- a/src/Raven.Studio/typescript/components/pages/resources/manageServer/certificates/partials/authEnabled/CertificatesCloneModal.tsx
+++ b/src/Raven.Studio/typescript/components/pages/resources/manageServer/certificates/partials/authEnabled/CertificatesCloneModal.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable react-compiler/react-compiler */
 import { yupResolver } from "@hookform/resolvers/yup";
 import genUtils from "common/generalUtils";
 import { FormInput, FormGroup, FormLabel } from "components/common/Form";