… src/Raven.Studio/package.json
… src/Raven.Studio/webpack.config.js